requirements.txt
Django==4.1.5
psycopg2==2.9.5
django-redis==5.2.0
redis==4.4.2
djangorestframework==3.14.0